refactor(nonces): lift nonce primitive to Features/Nonces, options as inheritance
Promotes the stateless nonce-service primitive out of the DPoP feature
folder so future consumers (state-parameter validation, anti-CSRF) can
reuse it without an obsolete-shim trajectory à la IJwtReplayCache.
INonceService, RollingHmacNonceService and friends move to
Features/Nonces; the cache-key prefix is no longer DPoP-bound.

Options become a small inheritance hierarchy: NonceOptions is the
generic base (acceptance window, rotation cadence); DPoPNonceOptions
extends it with the four per-endpoint require-nonce booleans, hosted
at OidcOptions.DPoP.Nonce. The active service still binds the DPoP
slot — one consumer today — but second consumers can subclass
NonceOptions independently. AddNonces() splits out from AddDPoP() so
the two registrations can be wired separately when needed.
