+++ Abblix.Oidc.Server/Common/Configuration/DPoPNonceOptions.cs
@@ -20,42 +20,33 @@
 // CONTACT: For license inquiries or permissions, contact Abblix LLP at
 // info@abblix.com
 
+using Abblix.Oidc.Server.Features.Nonces;
+
 namespace Abblix.Oidc.Server.Common.Configuration;
 
 /// <summary>
-/// Configuration for the DPoP-Nonce service (RFC 9449 §8 / §9). Governs the
-/// issuance window, secret-rotation cadence, and per-endpoint policy for
-/// requiring clients to include a server-issued <c>nonce</c> claim in their
-/// DPoP proofs.
+/// DPoP-specific extension of <see cref="NonceOptions"/> per RFC 9449 §8.
+/// Inherits the generic primitive's <c>AcceptanceWindow</c> and
+/// <c>RotationInterval</c> so DPoP can override them independently of any
+/// other nonce-service consumer, and adds the per-endpoint policy that says
+/// which OAuth endpoints reject requests whose DPoP proof omits the
+/// <c>nonce</c> claim (responding with <c>use_dpop_nonce</c> and a fresh
+/// <c>DPoP-Nonce</c> header).
 /// </summary>
 /// <remarks>
-/// The nonce service is stateless: nonces themselves are not stored. Only a
-/// short-lived rotating HMAC secret lives in <c>IDistributedCache</c>, keyed by
-/// time bucket so multiple server instances can validate each other's nonces
-/// without coordination. Per RFC 9449 §11.3 a nonce mismatch is recoverable —
-/// the client receives a fresh <c>DPoP-Nonce</c> header and retries — so the
-/// brief window during secret rotation where two instances disagree on the
-/// current secret degrades to a single client-side retry, not a hard failure.
+/// DPoP is the only nonce-service consumer at present, so the active
+/// <c>RollingHmacNonceService</c> binds directly against this type via
+/// <see cref="OidcOptions.DPoP"/>.<see cref="DPoPOptions.Nonce"/>; the
+/// inherited <see cref="NonceOptions.AcceptanceWindow"/> and
+/// <see cref="NonceOptions.RotationInterval"/> drive issuance, the DPoP-
+/// specific bools below are consumed by the token-endpoint integration
+/// slice. When a second consumer (e.g. state-parameter validation) arrives,
+/// it brings its own <see cref="NonceOptions"/>-derived options class and
+/// its own service registration; the inheritance hierarchy is what lets
+/// each consumer diverge on window and rotation independently.
 /// </remarks>
-public class NonceOptions
+public class DPoPNonceOptions : NonceOptions
 {
-    /// <summary>
-    /// Maximum age of a server-issued nonce that the validator will still
-    /// accept, measured against the timestamp embedded in the nonce. Defaults
-    /// to 5 minutes — long enough to survive normal client clock skew and a
-    /// retry round-trip, short enough that a leaked nonce stops being useful
-    /// quickly.
-    /// </summary>
-    public TimeSpan AcceptanceWindow { get; set; } = TimeSpan.FromMinutes(5);
-
-    /// <summary>
-    /// How often the HMAC secret used to sign nonces is rotated. Each rotation
-    /// boundary becomes a new cache bucket; <see cref="AcceptanceWindow"/>
-    /// MUST be larger than this so an in-flight nonce signed under the
-    /// previous bucket's secret is still verifiable. Defaults to 2 minutes.
-    /// </summary>
-    public TimeSpan RotationInterval { get; set; } = TimeSpan.FromMinutes(2);
-
     /// <summary>
     /// When <c>true</c>, the token endpoint rejects DPoP requests that omit a
     /// nonce claim with <c>use_dpop_nonce</c> per RFC 9449 §8. Defaults to
